refactor(react/Form): extract Input component
